Move non dom update functions out of domUpdates.js
